dev-requirements.txt
certifi==2025.6.15
click==8.2.1
coverage==7.8.2
dnspython==2.7.0
email-validator==2.2.0
fastapi-cli==0.0.7
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
iniconfig==2.1.0
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
packaging==25.0
pluggy==1.6.0
pygments==2.19.2
pytest==8.3.5
python-multipart==0.0.20
pyyaml==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
ruff==0.11.10
shellingham==1.5.4
typer==0.16.0
typing-extensions==4.13.2
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
